Make name, capped_size readonly in ORM and admin
Also prevent deletion of objects referenced by datasets

Because it would be a lot of work to be able to change these in Backdrop
right now - we're going "create only"

It should not be possible to delete data-groups and data-types used by data-sets.

We do validation in clean() to check that "readonly" fields haven't
changed.