Only create release tags for Python 2.7 job
Create the release tag if the Python 2.7 job succeeds. Don't do anything
in the Python 3 job.

This is a bit of a workaround for the fact that there's no way of
running scripts after *all* jobs complete on Travis. That means we end
up trying to release tag in every successful job.

See https://github.com/travis-ci/travis-ci/issues/929