Default to using PostgreSQL in development
Until now we've been using SQLITE as default (which has advantages, such
as being really easy to delete and access).

Constraints are enforced differently in PostgreSQL (ie at all) so we
need to start being able to use it ASAP, lest we run into problems on
preview that we haven't seen in development.

You can still use ``USE_SQLITE=true|false`` to switch the behaviour.