Increase DataSet name length from 50 to 200
We had a dataset in Backdrop which was 53 characters, so the import
script failed. This only because obvious when we started using
PostgreSQL in development.

Note that this will be our first schema migration that (should) happen
automatically on deploy, so we ought to check that the preview database
changes after we merge :)