Add dataset count to data group/data type admin lists.
When browsing data groups and data types, it's helpful to be
able to see how many data sets each one has associated with
it.

Add this to the Django admin list display for data groups and
data types.

This creates a sortable column in the admin, so users can sort
it to find the most and least popular data groups and data types.

Follows the approach used here:
http://stackoverflow.com/a/7448615/194000