Make sure that DataSet's query set deletes from Backdrop
Previously we had just implemented this in the model's delete() method, which meant that Backdrop would only be contacted when the user deleted the data set from the details page.

This change makes sure the model's delete() is called when delete() is called on the query set. This means that the delete action in the admin interface's list page will work.

https://www.pivotaltracker.com/story/show/70503022

[#70503022]