Purge cache in save() for both new and existing data-sets
Previously if users tried to access URLs for data-sets that did not exist, Varnish would cache the 404 responses, but when the data-set was later created in Stagecraft the cache would not be purged, and those cached 404 responses would continue to be served up. These cached responses are now purged when the data-set is created.

https://www.pivotaltracker.com/story/show/70022038

[#70022038]