Tests for copy data migration stuff, add client to requirements
These tests expect a method `migrate_data_set` which will take an old data set config, override specified parts of it and create a new data set with this specification (or update an existing one that matches this identifiers of the exising one). Then it will load the old data sets data using the backdrop client, convert the data to a new format defined by `data_mapping` and post it to the backdrop data-set for the newly created data set config.