Migration for the two mislabeled transactions by channel data sets
We are waiting to hear from cliff whether we also need to change the values for weekly claims to clerical_received being paper and so on. As monthly claims is not used we will not migrate this into the new carers allowance transaction by channel data set. There may be a requirement for other data sets to be combined in future however so we should determine if a script can add data to an existing data set already or if we will need to amend it. Adding to an existing data set with this script will require amending auto id so values are not overwritten.