Remove backdrop dependency for data set creation
Backdrop will create data sets on it's side as they're written to now
the dependency on backdrop can be removed from stagecraft for data set
creation.

This avoids the cross app transaction bugs around creating data sets.

The dependency has not been removed for data set deletion. It is already
a much more resilient process because if backdrop doesn't know a about a
data set stagecraft just caries on.