Add models and migration for Node and NodeType
This uses the django-mptt library to handle the database structure and
queries required to manage a tree.

It is worth noting that this will not cover absolutely every organisation
in government as it only supports a single parent. Very few organisations,
which luckily don't have any high volume services have multiple departmental
parents.

The organisation structure has otherwise been kept as simple as possible
with nodes being very abstract and identified with a defined type. This
should allow us flexibility as we layout our model of government.