Add Module and ModuleType to dashboard app
Dashboards' have many modules and we need to represent them. Each module
has a type which enforces the schema for the options JSON blob. This
JSON blob for options exists so that we do not need a table for each of
the module types. We are not going to be querying into this field as
part of an app, only ad-hoc so a JSON field should be perfect.

The two main requirements for modules are that no two modules can have
the same slug on the same dashboard and that a module's options should
validate against its type's schema.

Slug uniqueness test has a transaction.atomic wrapper on the create
because of the way django deals with transactions:
http://stackoverflow.com/questions/21458387/transactionmanagementerror-you-cant-execute-queries-until-the-end-of-the-atom