Allow empty query_parameters on add module
When adding a module to a dashboard, if there is no data set then allow
query_parameters to be present but empty rather than just not present.

The admin app will always provide query_parameters regardless of whether
there is a data set. It will provide an empty dictionary by default.