Lookup dashboards by slug
- Gives us nicer urls like /edit/name-of-dashboard
- Slug is also a unique identifier
- Avoids exposing uuid to end user

I prefer this way of looking things up because it's kind of human readable but at the same time works for the computer. UUIDs are still really useful for how we internally manage resources. For example we could potentially look up the history of a resource whose `slug` has changed multiple times, but the `UUID` will have remained the same.