Import schemas script
This has been tested on preview and locally against real data. All but
two modules are valid according to it and they are easy to fix. We
Should continue to only validate options as other flat fields are
already validated. There is an issue around the validation of the
availability module as it is always in the tab module so is not
currently validated.

Schemas are currently draft3 which necessitated some test and code
changes so these are not flagged as invalid.