Updated help text for seconds out of date
The way we calculate out of date data is based on the timestamp at the start of the period. The current defaults in stagecraft help text show the number of seconds of the period the data represents. This means the alerts start firing before the data is technically out of date. I have doubled the values so that the the correct values are entered when creating new datasets