Post create data set working
Currently disabled update as not in scope and may not be the best place
for it.

Validation for schema and actual existing group and type present

Includes some refactoring of the way we generate 400s

Clean up some of the fixture data to better represent actual data and
update references.