Increase size of datatype name field
This is due to running transform migrations. We want to create a
transform type for the module
g-cloud/proportion-of-sales-from-small-and-medium-enterprises
and use the module slug as the data type name. This is 53 characters,
but setting the limit to 60 chars as it is a round number and shouldnt
hit the limit on mongo collection names which is 128 characters