Route to get dashboards by tx service id
In order to map data from the transactions explorer dataset to the
relevant dashboard slug we need to be able to pass in a service id
and get back dashboards that use it.

The mapping from a service_id to a dashboard should be 1 to 1 as
dashboards are meant to represent services. There are unpublished
dashboards that use data for services for playing around/testing so we
filter these out to avoid making the wrong decision.