Create the dataset we are using
Transforms require the output dataset to exist before they can be
created. This creates in an idempotent way, so we can rerun the script
if necessary. Generate a bearer token - this code was stolen from a db
migration script - as this is also a throwaway script I didn't move it
somewhere more permanent