Run transform from latest non null data
Find the latest non null transform data and run the transform starting
then, with an end_at of 1 day later so the transform just picks up this
data point.

This wont work until we relax the latest data restriction in the
transform in backdrop, which prevents backfilling of data (which we are
trying to do here...)

Doctests were added for the add_day function - these pass on python 2.7
but not 3. I have skipped them because they are still useful, but as we
will never run this migration script on python 3, and it is a throwaway
migration script, it doesnt seem worth fixing to work with python 3

fixup! doctest