Creates transform type and transform for tx, runs it
When we run it we set the start to the earliest possible timestamp in
transactions explorer and the end at to significantly after the latest.
This ensure we will collect the latest data.

In doing this we realised the transform needs to ensure it gets the
latest non null value or null if none and posts it - currently it will
overwrite with nulls.

Tested and working on dev vm.