Wrap ResourceView POST in a transaction
We may have to save the model to the database before we have
successfully validated the resulting model (in the case of many to many
relationships as these need an id for the model to be able to add a
relation) and there is a possibility that the validation step may fail
and we would want to roll back the changes to the DB.

The tests for this are not the cleanest as we have to monkey patch on to
the Node model for these tests. This is because we need a model that is
part of our project and can be saved into the DB to ensure the rollback
takes place. I also wanted to ensure that we could cause failure rather
than abusing some fail case from a model.