Switch organisation tree to a graph
The change of the structure of the Node model broke the fixture for the
Node tests. I have switched these tests to use Factory Boy as that is
what we have adopted elsewhere. They set the id of all models to
maintain parity with the existing tests.

The dependency on the django-mptt library is still there as if removed
it will cause the migrations to fail. It can be dropped from the 2nd,
newly added migration but not from the initial. Frustrating, but can be
cleaned up one this has been rolled out to production

I have had to add a get_ancestors function to the Node model to maintain
the interface that was provided the MPTT model.

In NodeView.update_model we have to save() the model before we can add
any parents to it as django requires having an id to add the relation as
it needs to inject it into the join table. This is why I have added a
transaction around the POST method in ResourceView, so that if the view
method fails post update_model it will rollback the save here.

SQL in NodeManager
------------------

A `WITH` query is somewhat akin to a let statement and can be used to
create a labeled sub-query that you can reference in your main query. So
instead of:

```
SELECT *
FROM (SELECT a.foo, b.bar FROM a INNER JOIN b ON a.id = b.id)
WHERE foo > 10
```

you could say something like:

```
WITH sub_query_name AS (
  SELECT a.foo, b.bar
  FROM a INNER JOIN b ON a.id = b.id
)
SELECT * FROM sub_query_name WHERE foo > 10
```

This is especially useful if you want to reference the results of the
sub query multiple times in your main query.

http://www.postgresql.org/docs/9.3/static/queries-with.html

With the `RECURSIVE` you can turn a `WITH` query into something far more
powerful. You define and initial query to return a set of rows much as
you would do with a regular `WITH` query, but then you can define a
second query that is self referential and will be iterated on starting
with the self reference pointing at the result of the initial query.

An example is:

```
WITH RECURSIVE range(num) AS (
    VALUES (0)                        # sets initial num value to 0
  UNION
    SELECT (range.num + 1) AS num     # adds 1 to the previous num value
    FROM range                        # references previous iteration
    WHERE num < 10                    # limits the iteration to num < 10
)
SELECT num FROM range;
```

You will hopefully notice the `UNION` keyword between the initial and
interactive query, this joins together all of the result sets and drops
any duplicates. If you were to add `UNION ALL` then it would include
duplicates.

Hopefully this will give you some introduction to the SQL being used in
this commit. There is some code that switches out the initial query so
that we can either start with the node that we are finding the ancestors
of or start at the first level of ancestors.