Take spreadsheet merging output and write redirects
Skips redirects when no existing page. This is done by checking the page
returns a 200. This seems potentially flakey so if the status code is
not a 404 or 301 we raise an error - this url needs further
investigation. If this is too problematic we can find out how
transactions explorer determines whether to create pages and apply
the same logic.