Allow 'dashboard-editor' users to see all datasets
Allow users with the common dashboard-editor role to see all datasets. This is
required for the admin app to show users a list of dataset groups and
types in a dropdown list.

Although this is part of the base resource view, this should not affect
endpoints other than datasets as they are the only models that have the
backdrop_user relationship which is used for filtering.