Add dogslow logging
A Django middleware that logs tracebacks of slow requests.

https://pypi.python.org/pypi/dogslow

I have configured this to log to the stagecraft.apps logger - ideally we
should setup a new logger for this, but its not immediately obvious how
to do this.

Unfortunately the timeout is only whole seconds - hopefully setting this
to 1 second wont add too much noise to our logging, but we can monitor
it and disable if so.