Add django admin for transforms
At the moment there is little visibility into what transforms exist and
how you would go about configuring and creating them.

The __str__ function in TransformType has been added so that the name of
the transform type appears in django admin

I've added to the list display fields as the input/output groups are
important to understand what the transform is operating on.