Link more things to their parents
There are some problems though:

Having things twice in org dict means that it will create the wrong type
of node if it is not the matched node (as the link to parent also
corrects the type) and the not matched version is created second (it
will override the type in this case).

The test's are failing. In the last test there is not ' on attorney
general's office in the key, however there is on the full org dict
(including services and transactions)

For some reason the ancestors created are wrong....