Improve organisation and dashboard loading/querying
This improves the organisation loading code and
updates it against the latest spreadsheet loading code.

It also improves dashboard loading and increases the
length of the title field.

Denormalised fields have been added to dashboards to
improve the loading of /public/dashboards.

This also fixes query parameters for the quarterly
and digital takeup modules.

Django emulates CASCADE which will delete dashboards associated
to organisations when the the organisation is deleted.

Prevent this from happening by setting foreign keys to NULL
beforehand.