Add a request-id to all log messages
For tracing the 502 errors we are seeing.

From kibana, we can trace an nginx 502 with a request-id to a single log
line in stagecraft, which is when the request processing is started.
From here we can get a process_id, but it is impossible to work out
which log lines refer to the same request as a process can handle
multiple requests.

This uses the idea in
https://raw.githubusercontent.com/jedie/django-tools/master/django_tools/middlewares/ThreadLocal.py
to store the request in a thread local variable, which we can then
reference when logging