Fix module creation
This attribute `--update`` now *only* deals with updating dashboard
attributes. The default is for the script to update any dashboards
attributes that are not `high_volume`

If `--update` is passed, then `high_volume` dashboards will also be
updated.

This change was introduced so that we always update modules when we are
creating a dashboard or if the dashboard does not have any modules -
previously this would not do this for high_volume dashboards unless
`--update-all`` was passed in.