Include Agencies even when missing an abbreviation.
If we've had issues with Agencies that don't have an abbreviation missing from the org graph.
This means that we have been mis-representing the structure of government.

If an abbreviation doesn't exist for an agency, still record the agency details against
the department. Set the slug to the agency name.