Add more exceptions in TX -> GOV.UK mapping
A couple of organisations have changed names recently so we have
inconsistency between the TX spreadsheet and the names in GOV.UK. As
well as these, fixing the dropping of agencies with no abbr was brought
up a couple of other mappings to be made.

'research councils' is a fake agency grouping together the many councils
inside of BIS, these should be associated we BIS as they can't be broken
up to belong to any one of the child councils

'planning portal / department' had to be a starts with because the
string used was painfully long and this is unique enough to pull out the
one use in a service DCLG provide.