Multiple ids for resource view
We want to change how we express the IDs that a client can use to access
a resource. This is to include the case where we have two different ways
of referencing a resource (id, slug) used in dashboards.

This moves all of the logic about how the resource is expressed into a
single place, the view sub class rather than being split across urls.py
and the view.