Add PUT method to resource view
The motivation for this is to be better HTTP citizens and update models
using PUT rather than POST. This also lets us get rid of get_or_create,
which has caused some issues when trying to change how we parsed ids.

None of the APIs exposed using ResourceView have clients relying on the
fact that POST also updates (this is only dashboards), so change the API
and removing the checks on 405 response for PUT should be fine (famous
last words, something about eating my hat).

ModuleView was expecting PUT to be blocked as it doesn't currently have
all of the validation in place to be able to add or edit a module
through its exposed API.