Breakup updating relationships from model fields
We can only update the relationships to a model once it has been saved,
so it makes sense to split this out from update_model to after it has
been saved.

This was previously hacked into update_model for nodes when we wanted to
update the parent, but we now also want to do this for dashboards so it
makes sense to do this in a slightly more pleasant way.

Have smooshed together validate and save as it was been repeated in a
few places and starting to look a little verbose.