Create a new Dashboard View that inherits from Resource View.
Change the Module View to inherit from Resource View.

Clean up deprecated functions to add a module to a dashboard.
This is now being done by the update model function.

Add status to the fields returned by the view serialize function.
It is was originally being output by the model's serialize and
is used by other applications.

The Performance Platform Admin App expects the Organisation to be
returned as part of the Dashboard data as an object containing
the name and uuid, not just the uuid.

When Dashboards are updated, they should only pass back the id
of the organisation in the create and update request, not a
whole organisation object.

Resource view returns errors as strings not json objects.
Dashboards and modules therefore now return errors as strings.

Updated one test to check that an error was returned rather than
the format of the error.  It will be a much larger task to update
Resource View and all of the views that inherit from it to return
errors in json format.