Fix tests:
1. get_fields_with_model method deprecated and not backwards
compatible with django 1.7

The values returned by the get_fields and get_fields_with_model
functions have changed in django 1.8.

If the app name is included in INSTALLED_APPS, the functions now
also return relationships in the list of fields. The django
documentation advises excluding relationship fields when getting
the list of fields:

https://docs.djangoproject.com/en/1.8/ref/models/meta/#retrieving-all-field-instances-of-a-model

Switched to using get_fields as get_fields_with_model has been
deprecated.

2. Set default value of options field in Module and schema field in
ModuleType to an empty dictionary.

The database fields are not nullable and this was causing
JSONField to throw a validation error.

3. Fixtures weren't loading due to presence of class level setUp
and tearDown methods.

Changed failing tests to not use fixtures.

4. Missing tearDownClass methods.

The absence of tearDownClass was causing tests to fail.

6. Executing the QuerySet delete method does not call the delete
method on individual objects.

Explicitly call the delete method on individual objects to test
that the  delete method on the backdrop API is being called.