Remove datasets annotation.
In Django admin, the lists for Users, Data Groups and Data Types
is sorted by the number of data sets the model "has" or for Users,
is authorised to edit.

This was done by annotating the queryset with the count of datasets.
For reasons unknown this doesn't work in django 1.8 and has been
removed.