Make the creator of a resource the owner and enforce owner-only access
When a resource is created, the user creating it is added to the list of
owners. In the case of dashboards, datasets and transforms the list is
associated to the model directly.

In the case of modules, ownership is delegated to the dashboard the
module is associated with.

The owner is the only non-privileged user who can view and modify their
resources.

This adds various tests and improves testing in places where it was
lacking. Specifically, ownership interactions across resources is
tested, and transforms are better tested.