Consolidate roles and permissions with Admin App
 - Introduced admin and dashboard-editor roles and permissions
 - The admin role can do anything and can see all dashboards
 regardless of whether they are an owner of the dashboard
 - The dashboard-editor role has all the permissions necessary
 to use the features accessible through the Edit dashboard button
 but can only see dashboards for which the user is an owner

The dashboard-editor has been granted a POST permission for
Dashboard resources so as to allow users with this role
to add modules to their dashboards. This was necessary
because the API for adding a module to a dashboard hits the
Dashboard POST route before being passed through to the
Module POST route.