Add delete method in resource view.
Add delete method to delete unpublished dashboards based on model attribute as want to ignore in
review status in stagecraft.  Also cannot delete models without published attribute as want to delete
the dashboard and not modules specifically in case the delete cascades.

Add tests in resource view to test that unpublished dashboard can be deleted, published dashboard cannot,
and a module without the published dashboard cannot be deleted.