Updating module query_params to untruncated slug
We believe it would be safe to do this for non truncated but it's
probably better to limit it and then perhaps take out the truncated slug
related code after all the data is migrated. As a result the test is
pretty long - it should be okay if we will tidy this up anyway.