Create digital takeup module even when value 0
The summaries page will no show these 0s so not having a module on the
dashboard is confusing. This seems to have originally been prevent
deliberately though as the other modules have a not None check. Is
digital takeup a special case? If so the real change should be changing
the transform to ignore only 0s from digital takeup. It's worth noting
that the totals on the services page also ignore 0s (also for digital
takeup?) again. Was this deliberate?