Collectors models and tests
We have built the underlying data model for collectors in the
Performance Platform.

https://drive.google.com/open?id=0B4BcZ31U-tWQVjZmQVd4UUNXT0t2bXQyQk9pR0RGWUpDUGF3

This uses Fernet to encrypt the credentials we store for 3rd party
services. This is to protect data that has a reasonable impact potential
if leaked when the data is at rest. Fernet uses AES and HMAC under the
hood and should provide adequate protection for our data.

Along side all of the schema tests for the options/query/credentials we
enforce some ownership constraints and some provider constraints. We
want to be sure that a collector uses a type and data source for the
same provider, so we don't end up with a GA collector talking to a Piwik
source. We also want to be sure that the user interacting with the
Collector is associated with both the data set and the data source, so
they don't end up using someones random data source and pulling their
juicy data.

Because of the constraint on providers needing to be the same between
the type and data source, when creating a collector through a factory we
have to pull the provider from the type into the data source. If we
didn't do this then it would create two different providers for the type
and the data source

We are relying on a git tag of django-field-cryptography as this was not
released into pypi correctly and we need some of the features - ie to de-encrypt
the data returned from the database.  We are using a specific version of
cryptography in the requirements:

https://github.com/pyca/cryptography/blob/0.9.3/src/cryptography/fernet.py#L84

This change correctly identifies invalid tokens, whereas 0.6.* would fail
when data was falsy.