Create collector type schemas
1. Add schemas for the different types of collector.
Each collector type needs three schemas:
* query.json - used to validate the query parameters of the collector
* options.json - used the validate the optional parameters of the collector
* descriptor.json - describes the properties required to create a collector type.
Used by the import_collector_types script to extract properties e.g. The title in the
descriptor schema will become the human readable name of the collector type.

Every collector type should have a query, options, and descriptor schema.
Even if a collector type doesn't allow for a query and/or options object, a
schema with empty properties has been added to be able to validate that no
properties can be passed in.

All of the schemas have additionalProperties set to false so that we can
catch any collector type attributes that may have been missed when the
collectors are imported.