Script to create collector types.
The folder names that contain the schemas for a collector type follow a pattern.
If the collector in performanceplatform-collector can be found in
performanceplatform/collector/ga/trending then the folder name for the schemas should
be 'ga-trending'

The collector type slug is determined by the name of the folder containing
the schemas for the collector type. e.g in the example above the collector type slug
is 'ga-trending'

The provider is determined by trying to match the name to the first part of
collector type slug. e.g. if the slug is 'ga-trending', the provider name is 'ga'.
The provider is created if it doesn't already exist.

Provider creation should be revisited to perhaps add a slug to the Provider and base Provider
selection on the value of the slug.