Load the schemata used by the collectors app
We want a script that will import the schemas used for collector types
using the ones created in the 2nd last commit and that now live in
collectors/schemas/collector-types.

This will use these files to find the correct CollectorType model and
then save the query and options schemata on it. This will validate the
schemas to make sure they are correct before saving and then validate
associated collectors.