Import collector configuration from repo
Currently all of our collector configuration is stored in a git
repository, this will parse all the queries and store them in the
database given the path to the root of repo.

The entrypoint_information dict is based off the value in
performanceplatform-collector-config/tools/cronjobs.py to try and have
parity with how collectors are running at the moment.

In order to identify each individual collector we generate a slug for
it. This is composed of the data group and type that the collector will
write into, but as there are sometimes more than one collector that
writes into the same data set we need to go further. We hash the query
and options as there should be no collector with identical values as
they would be collecting the same data. (There actually were a couple,
which I have been back and removed)