Check that collector models can be listed.
Models that have an owners attribute need to have a for_user method
defined in the model manager to be able to call list on the models.

Added a Manager for the Collector and DataSource models.

Added tests to check that all of the associated collector models can
be listed if the user is an admin or has the collector permission.