Remove old scripts
These were a collection of scripts used to bootstrap various bits of our
system as we created new features. These have been used and are now no
longer needed, so they should be dropped from the repo. They will always
be in the history if people want to find and reference them.