Breakup backdrop URLs so we can talk internally
Previously, when we were on our own infrastructure, we had internal DNS
so that we could use the public facing urls for services and they would
route through to the internal services without going out of our stack.
Now, we need to explicitly talk to the internal services so we don't go
out to our CDN and back into our stack. We also now have different
internal URLs for the read and write instances of backdrop, rather than
one that splits based on the HTTP method used (GET -> read, POST/PUT ->
write).

This PR breaks up the use of backdrop URLs so that we use the correct
interal URL depending on the context of use (getting data or mutating
data). This also includes a public url so that the admin interface can
display a URL for the dataset that a consumer of the platform can see.