Add empty data set button to data set form
At the moment for an administrator to empty a data set on the
performance platform they have use curl to do it. It requires pulling
some information from the django admin interface and then remembering
the format for the command. This is less than ideal, so we've switched
this for a button on the data set form that will let you dump the data.

This works by overriding the template used by the DataSetAdmin to render
the edit form. Django uses values on the submit button to indicate
different kinds of action to be taken on a model. This has introduced a
new one and overrided the handler to intercept this and call backdrop to
dump the data set.