Create celery app and register the task.
1. The new run_collector task is not being registered in celery because
shared tasks need a celery app to bind themselves to.

2. Allow Celery to auto discover tasks.

3. Problems with how credentials are stored/parsed/handled:
If the credentials are saved correctly we think this will work. Currently
the are not being decrypted when loaded. This is because the key to decrypt
is environment specific and also copied down (e.g. preview and development
cannot decrypt production credentials).

Once the credentialss are in the right format we get this error for GA:
  InvalidClientSecretsError: File not found: "/data/apps/performanceplatform

We do not yet support non file based creds in the GA realtime and trending
collectors. We need to support passing the whole files in rather than
just paths to files.  The core  GA collector should work as this allows for
the credentials to contain different key/values, however we have been unable
to confirm this.

We need to discuss how we're going to handle having different credentials
in the database per environment.

Updating the collectors to allow for real credentials rather than a path to a
file will be handled in another story/pull request.