Expose OAuth 2.0 credentials via a Storage Object
The Google Analytics collector needs special treatment. We can't
just store the encrypted credentials in the database and then
pass to the collector as needed.

Google Analytics requires that you have application specfic (e.g.
for stagecraft) client secrets and a user specific token. Those are
stored together in a Storage Object.

We created our own Storage Object rather than use the one created
for Django to allow us to encrypt the credentials in the Data Source
model.