Fix backfill collector bug.
Passing start and end dates to the run collector task caused it to
fail with a json serialization error as json does not support dates.

The code to convert the start and end dates from a string to a date
has been moved from the view to the task itself, thereby avoiding
the problem.

Added checks to the run_collector controller to make sure that the
date string being passed in is in the correct date format.

We didn't pick this up earlier as the post data is only
serialized when it's being added to the celery queue, and not when
the task method is called directly.