Disable delete button on dataset if dashboard is published.
If the dataset is linked to a dashboard that is published, disable
the delete data set button that appears at the bottom of change page
in django admin.

Display list of linked published dashboards

For data sets where the delete button is disabled because they are
linked to a published dashboard, display a list of these dashboards.