Add a check for data sets linked to tabbed modules.
Tabbed modules are child modules of another module. Rather than be
linked as a child however, they are configured in options as a list
of tabs.

Add a property on data sets that returns a list of the modules and
tabbed modules associated with the data set.

Use the module property in data set admin to find linked dashboards.

Also created a list of data sets that should never be deleted. This
is because they are linked to so many modules (>2000) that it was
causing the django admin change screen to return a 502.