Make Link model editable in Django Admin
Other links, to things like policies and publications are not
editable in performanceplatform-admin. It seems that these links were
added to some of the site activity dashboards in the early days of
performance platform, when the config for dashboards was stored in as
json files in the spotlight-config repo.

The contents of spotlight-config was then migrated to stagecraft using the
spotlight_config_migration script. This script is the only place I could find
where links are actively added to dashboards.

It is possible to add other links to a dashboard by using the stagecraft api,
but as they are not included in the post data from performanceplatform-admin,
it can be assumed that these links are no longer supported.

A zendesk ticket was raised recently because to the link to the Environment
Agency's policies are incorrect. We need a way to update any other links a
dashboard may have to fix the zendesk ticket, before deciding if it is worth
adding extra functionality to the admin/self-serve app.

Zendesk ticket: 1190780