Remove circular imports
Changed Node and NodeType to call a serialize method in their
models rather than their Views.
The dependencies between Dashboard and Module means that the module property
in DataSet was trying to find an authenticated user, eventhough the code
is being called from the model layer.

Also fixed a circular import that has been introducted between DataSet
and Module.