Service ID missing from backdrop data causes import to fail
Some dashboards are missing the service_id field in the data
retrieved from backdrop.

The missing field is in the digital_takeup data for all of the
dashboards.

As far as we can tell, the digital_takeup data has not been updated
since July 2015, though the dashboards on the frontend show data
through to December 2015.