Abbreviations longer than 50 characters cause import to fail
Abbreviations longer than 50 characters in length cause the
import_organisations script to fail.

In the Node (organisation) model, the abbreviation field is limited
to 50 characters.

When the model is serialized or made ready for spotlight, if an
abbreviation does not exist, the name is passed back instead.
The name field allows upto 256 characters.

Have updated the import_organisations script until it is determined
whether of not it is safe to update the Node model.

Pivotal: https://www.pivotaltracker.com/story/show/110412434